% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_throws_percent.R
\name{free_throws_percent}
\alias{free_throws_percent}
\title{The Free Throw Percentage}
\usage{
free_throws_percent(year, player, data)
}
\arguments{
\item{year}{numeric value, the year sought for free throw percentage}

\item{player}{character value, the player that produced the free throw
percentage}

\item{data}{dataframe, the dataset that is analyzed}
}
\value{
A numeric value that shows the free throw percentage
}
\description{
The Free Throw Percentage
}
\examples{
nba <- read.csv(file = "../data/Seasons_Stats_NBA.csv")
free_throw_percent(year = 1985, player = "Mike Mitchell", data = nba)

}
