% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corr_plot.R
\name{make_corr_plot}
\alias{make_corr_plot}
\title{Makes a Correlation Plot}
\usage{
make_corr_plot(data, year)
}
\arguments{
\item{data}{A dataframe, the data for which you want to produce a
correlation plot}

\item{year}{A numeric value, the year for which you want the data}
}
\value{
A plot that shows the numerical value of correlations between
numeric variables in the data
}
\description{
Makes a Correlation Plot
}
\examples{
nba <- read.csv(file = "../data/Seasons_Stats_NBA.csv")
make_corr_plot(data = nba, year = 1985)

}
