% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player in a Year}
\usage{
oldest_player(year, data)
}
\arguments{
\item{year}{numeric value, the year that is intended to find the oldest player}

\item{data}{dataframe, the data set that is analyzed}
}
\value{
A character object of the name of the oldest player for the given
year in the dataset
}
\description{
Oldest Player in a Year
}
\examples{
nba <- read.csv(file = "../data/Seasons_Stats_NBA.csv")
oldest_player(year = 1985, data = nba)

}
